dockerfile_requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==19.3.0
bcrypt==3.2.0
blessings==1.7
cachetools==4.2.4
certifi==2018.1.18
cffi==1.14.5
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==1.2.2
contextvars==2.4
cryptography==3.4.7
cycler==0.10.0
dataclasses
deepdiff==5.7.0
dill==0.3.3
dm-env==1.5
dm-env-rpc==1.1.0
dm-tree==0.1.6
docker-pycreds==0.4.0
enum34==1.1.10
et-xmlfile==1.1.0
filelock==3.0.12
frozenlist==1.2.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.20
google-api-core==2.7.1
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.0
gpustat==0.6.0
grpcio==1.44.0
gym==0.15.4
h5py==2.10.0
idna==2.6
idna-ssl==1.1.0
immutabledict==2.2.1
immutables==0.16
importlib-metadata==0.23
inflection==0.5.1
jdcal==1.4.1
Jinja2==2.10.2
joblib==0.14.1
jsonschema==4.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
lz4==3.1.10
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib
mock==4.0.3
more-itertools==7.2.0
mpyq==0.2.5
msgpack==1.0.0
msgpack-numpy==0.4.7.1
multidict==5.2.0
multiprocess==0.70.11.1
numpy
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
open-spiel==1.1.0
opencensus==0.8.0
opencensus-context==0.1.2
opencv-python==4.6.0.66
openpyxl==3.0.5
opt-einsum==3.3.0
ordered-set==4.0.2
packaging==20.3
pandas
paramiko==2.7.1
pathtools==0.1.2
pdoc3==0.7.1
Pillow==6.2.1
pluggy
portpicker==1.5.2
posix-ipc==1.0.5
prometheus-client==0.13.1
promise==2.3
protobuf==3.19.4
psutil==5.6.5
py
py-spy==0.3.11
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
pydantic==1.9.0
pygame==2.0.1
pyglet==1.3.2
PyNaCl==1.4.0
pyparsing==2.4.4
pyrsistent==0.18.0
PySC2==4.0.0
python-dateutil==2.8.1
python-redis-lock==3.3.1
pytz==2020.1
pyxdg==0.25
PyYAML==5.1.2
pyzmq==19.0.1
redis==3.5.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
s2clientprotocol==5.0.10.88500.0
s2protocol==5.0.10.88500.0
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.1
SecretStorage==2.3.1
sentry-sdk==1.5.8
setproctitle==1.2.2
shortuuid==1.0.8
six==1.13.0
sk-video==1.1.10
SMAC @ git+https://github.com/oxwhirl/smac.git@845b1173c8e38b37f4c67c5a03c6614916fb2f38
smmap==5.0.0
tabulate==0.8.9
tenacity==6.0.0
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorboardX
termcolor==1.1.0
threadpoolctl==3.1.0
thrift==0.13.0
tqdm==4.38.0
trueskill==0.4.5
typing_extensions==4.1.1
urllib3==1.22
wandb==0.12.11
wcwidth==0.1.7
websocket-client==1.3.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.7.2
yaspin==2.1.0
zipp==0.6.0
